# docs/api/User_Api_Swagger.yaml
...
openapi: 3.0.0
info:
  title: ESP RainMaker Neo Auth & Accounts API
  version: 0.8.0
  description: |
    Sign-in and account management for the ESP RainMaker Neo cloud.
    Anything that produces or rotates an access token lives here; once
    you have a token, the rest of the platform lives in the
    **ESP RainMaker Neo Platform API**

    ESP User is the platform's own OAuth 2.1 / OIDC issuer: every token a client
    holds is minted here, regardless of how the user proved their identity.
    Admins authenticate directly against their AWS Cognito user pool.

    Login methods (all end in OUR tokens; upstream credentials never reach clients):

    - **Federated login (default)**: `GET /oauth2/authorize` starts a browser
      authorization-code + PKCE flow. With a single enabled identity provider the
      user is auto-redirected to it; the
      broker exchanges the upstream code server-side and issues OUR authorization
      code at the client's `redirect_uri`.
    - **Native password APIs** (`/v1/user/auth/*`): the main-era
      native surface. Credentials are verified against the bundled Cognito pool server-side and converted to our tokens.

    Token lifecycle after login:

    1. `POST /oauth2/token` (`grant_type=refresh_token`) - rotate the token set
       (native `/v1/user/auth/token/refresh` is reuse-tolerant instead).
    2. Send `access_token` as a `Bearer` `Authorization` header on subsequent
...
       must come from the same sign-in.
...
  contact:
...
  license:
...

...
servers:
...
paths:
...
  /oauth2/userinfo:
    get:
      summary: OIDC UserInfo endpoint
      description: |
        Returns the profile claims for the presented access token: `sub` always, `email` with the `email` or `profile` scope, `phone_number` with `phone`.
      tags:
        - OAuth2 / OIDC Core
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The subject's scope-authorized claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserinfoResponse"
        "401":
          description: |
            invalid_token — the access token is missing, expired, or invalid.
            Carries a `WWW-Authenticate: Bearer error="invalid_token"` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
    post:
      summary: OIDC UserInfo endpoint (POST variant)
      description: |
        Same as `GET`. The access token may also be sent as the `access_token` field of a form body.
      tags:
        - OAuth2 / OIDC Core
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                access_token:
                  type: string
                  description:
                    The access token, when not sent in the Authorization header.
                  example: "eyJhbGciOiJSUzI1Ni..."
      responses:
        "200":
          description: The subject's scope-authorized claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserinfoResponse"
        "401":
          description:
            invalid_token — the access token is missing, expired, or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
...
components:
  schemas:
...
    OAuthError:
      type: object
      description:
        OAuth 2.0 error response. Distinct from the API's generic `{message}`
        errors.
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - invalid_target
            - invalid_token
          example: "invalid_grant"
        error_description:
          type: string
          description:
            Human-readable detail. Does not distinguish why the token was
            rejected.
          example: "The refresh token is invalid, expired, or revoked."
...
    UserinfoResponse:
      type: object
      description: |
        UserInfo claims. `sub` is always present; `email` / `phone_number`
        appear only when the access token's scope allows them and the user has them set.
      required:
        - sub
      properties:
        sub:
          type: string
          description: The user's immutable user id (the token subject).
          example: "9aXb7cD..."
        email:
          type: string
          description: Present only with `email` / `profile` scope.
          example: "user@example.com"
        phone_number:
          type: string
          description: Present only with `phone` scope.
          example: "+15551234567"
...
  securitySchemes:
    CognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      description: |
        Admin Cognito User Pool JWT, for the admin APIs (`/v1/admin/*`).
        End-user routes take an ESP User token instead — see `bearerAuth`.
      x-amazon-apigateway-authtype: cognito_user_pools
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An ESP User access token, sent as a Bearer token. The token is verified
        against the provider's published public keys.
    clientBasicAuth:
      type: http
      scheme: basic
      description: |
        OAuth client authentication (RFC 6749 §2.3.1) as HTTP Basic
        `base64(client_id:client_secret)`. A confidential client sends its secret;
        a public client sends its client_id with an empty secret.
security: []